requirements.txt
fastapi
uvicorn
pydicom
Pillow
python-multipart
